+++ lua/mappings.lua
@@ -69,3 +69,7 @@ map("n", neorg_leader .. "t", "<cmd>Neorg toggle-concealer<CR>", { desc = "Toggl
 map("n", neorg_leader .. "h", "<cmd>Neorg keybind norg core.norg.qol.todo_items.todo.task_toggle<CR>", { desc = "Toggle TODO state" })
 map("n", neorg_leader .. "c", "<cmd>Neorg toggle-concealer<CR>", { desc = "Toggle concealer visibility" })
 
+-- live server 
+map("n", "<leader>ls", ":LiveServerStart<CR>", { noremap = true, silent = true }) -- Start the server
+map("n", "<leader>le", ":LiveServerStop<CR>", { noremap = true, silent = true })  -- Stop the server
+map("n", "<leader>lt", ":LiveServerToggle<CR>", { noremap = true, silent = true }) -- Toggle Server State